Promote CompositeValidationException to top level
